reduced to header-navigation
