+++ src/components/m-header-main-navigation-mobile/index.js
@@ -18,7 +18,7 @@ class AXAMainNavigationMobile extends BaseComponentGlobal {
 
     const offCanvas = getAttribute(this, 'offcanvas');
 
-    this.className = classnames(this.initialClassName, 'm-header-main-navigation-mobile', {
+    this.className = classnames(this.initialClassName, 'm-header-navigation-mobile', {
       'm-header-main-navigation-mobile--off-canvas': !offCanvas,
     });
