--- src/components/m-header-main-navigation/index.scss
@@ -1,191 +1,0 @@
-@import "settings/header-main-navigation";
-@import "../a-link/mixins/link";
-@import "../../styles/settings/colors";
-@import "../../styles/mixins/clearfix";
-@import "../../styles/mixins/button-reset";
-@import "../../styles/mixins/grid";
-@import "../../styles/mixins/respond";
-@import "../../styles/mixins/typo";
-@import "../../styles/mixins/unstyle-list";
-@import "../../styles/mixins/word-break";
-@import "styles/stroke";
-
-.m-header-main-navigation {
-  position: relative;
-  z-index: $header-main-navigation-z-index;
-
-  display: block;
-  height: $header-main-navigation-height;
-  border-bottom: $header-main-navigation-border-bottom-width solid $header-main-navigation-border-bottom-color;
-  background: $header-main-navigation-bg;
-}
-
-.m-header-main-navigation__box {
-  @include make-container();
-  @include make-container-max-widths();
-  @include make-row(null);
-
-  height: 100%;
-
-  flex-wrap: nowrap;
-
-  align-items: center;
-
-  @include respond-down(md) {
-    max-width: none !important;
-  }
-}
-
-.m-header-main-navigation__logo {
-  @include make-col(2);
-
-  align-self: center;
-
-  @include respond-up(sm) {
-    align-self: flex-start;
-  }
-}
-
-.m-header-main-navigation__logo-link {
-  display: block;
-
-  width: $header-main-navigation__logo-size;
-
-  @include respond-up(sm) {
-    width: $header-main-navigation__logo-size--md;
-  }
-
-  @include respond-up(xl) {
-    width: $header-main-navigation__logo-size--xl;
-  }
-}
-
-.m-header-main-navigation__logo-img {
-  display: block;
-
-  width: $header-main-navigation__logo-size;
-  height: $header-main-navigation__logo-size;
-
-  @include respond-up(sm) {
-    width: $header-main-navigation__logo-size--md;
-    height: $header-main-navigation__logo-size--md;
-  }
-
-  @include respond-up(xl) {
-    width: $header-main-navigation__logo-size--xl;
-    height: $header-main-navigation__logo-size--xl;
-  }
-}
-
-.m-header-main-navigation__nav {
-  display: block;
-
-  margin-left: auto;
-
-  @include respond-down(md) {
-    display: none;
-  }
-}
-
-.m-header-main-navigation__list {
-  @include unstyle-list();
-
-  display: flex;
-}
-
-.m-header-main-navigation__list-item {
-  display: block;
-
-  float: left;
-
-  margin: 0 $header-main-navigation__list-item-x-margin;
-}
-
-.m-header-main-navigation__list-link {
-  @include link-color();
-  @include typo-13--semibold();
-
-  position: relative;
-
-  display: flex;
-  height: $header-main-navigation-height;
-  padding: 0 $header-main-navigation__list-link-x-padding;
-
-  align-items: center;
-  text-align: center;
-  text-transform: uppercase;
-  vertical-align: middle;
-
-  &:hover,
-  &:active,
-  &:focus {
-    text-decoration: none;
-  }
-}
-
-.m-header-main-navigation__form {
-  display: block;
-  padding: $header-main-navigation__form-y-padding 0 $header-main-navigation__form-y-padding $header-main-navigation__form-padding-side;
-  margin-left: $header-main-navigation__form-margin-side;
-
-  border-left: $header-main-navigation__form-border-side-width solid $header-main-navigation__form-border-side-color;
-
-  @include respond-down(md) {
-    padding-left: 0;
-    padding-right: $header-main-navigation__form-padding-side;
-    margin-left: auto;
-
-    border-left: none;
-    border-right: $header-main-navigation__form-border-side-width solid $header-main-navigation__form-border-side-color;
-  }
-}
-
-.m-header-main-navigation__search {
-  display: none;
-}
-
-.m-header-main-navigation__search-icon {
-  display: block;
-  width: $header-main-navigation__icon-size;
-  height: $header-main-navigation__icon-size;
-
-  color: $header-main-navigation__search-icon-fill;
-}
-
-.m-header-main-navigation__burger {
-  @include button-reset();
-
-  display: none;
-
-  margin-left: $header-main-navigation__burger-margin-left;
-
-  color: $header-main-navigation__burger-color;
-
-  @include respond-down(md) {
-    display: block;
-  }
-}
-
-.m-header-main-navigation__burger-icon {
-  display: block;
-  width: $header-main-navigation__icon-size;
-  height: $header-main-navigation__icon-size;
-
-  fill: currentColor;
-
-  .is-burger-open & {
-    display: none;
-  }
-}
-
-.m-header-main-navigation__close-icon {
-  display: none;
-  width: $header-main-navigation__icon-size;
-  height: $header-main-navigation__icon-size;
-
-  fill: currentColor;
-
-  .is-burger-open & {
-    display: block;
-  }
-}
