--- src/components/m-header-main-navigation/settings/_header-main-navigation.scss
@@ -1,29 +1,0 @@
-@import "../../../styles/settings/colors";
-@import "../../../styles/settings/zindex";
-
-$header-main-navigation-bg: $color-prim-white !default;
-$header-main-navigation-height: 75px !default;
-$header-main-navigation-z-index: $zindex-header-main-navigation !default;
-
-$header-main-navigation-border-bottom-color: $color-prim-gray-mercury !default;
-$header-main-navigation-border-bottom-width: 2px !default;
-
-$header-main-navigation__logo-size: 35px !default;
-$header-main-navigation__logo-size--md: 43px !default;
-$header-main-navigation__logo-size--xl: 53px !default;
-
-$header-main-navigation__list-item-x-margin: 8px !default;
-
-$header-main-navigation__list-link-x-padding: 10px !default;
-
-$header-main-navigation__form-y-padding: 6px !default;
-$header-main-navigation__form-padding-side: 20px !default;
-$header-main-navigation__form-border-side-width: 1px !default;
-$header-main-navigation__form-border-side-color: $color-prim-gray-silver !default;
-$header-main-navigation__form-margin-side: 20px !default;
-
-$header-main-navigation__icon-size: 25px !default;
-$header-main-navigation__search-icon-fill: $color-prim-gray !default;
-
-$header-main-navigation__burger-color: $color-axa-blue !default;
-$header-main-navigation__burger-margin-left: 15px !default;
