+++ src/components/m-header-main/_preview.html
@@ -1,6 +1,13 @@
 <axa-header-main first-left>
   <axa-header-logo></axa-header-logo>
 
+  <axa-header-navigation items='[
+      {"url": "#", "name": "car insurance"},
+      {"url": "#", "name": "home insurance"},
+      {"url": "#", "name": "health insurance<br>sub line"},
+      {"url": "#", "name": "travel insurance"}
+    ]'></axa-header-navigation>
+
   <axa-button>foo</axa-button>
   <axa-header-search></axa-header-search>
   <axa-header-burger></axa-header-burger>
