+++ src/components/m-header-navigation/_preview.html
@@ -1,8 +1,6 @@
-<axa-header-main-navigation items='[
+<axa-header-navigation items='[
   {"url": "#", "name": "car insurance"},
   {"url": "#", "name": "home insurance"},
   {"url": "#", "name": "health insurance<br>sub line"},
   {"url": "#", "name": "travel insurance"}
-]'>
-  <a href="#" class="m-button m-button--ghost m-button--motion hide-md-down">Pay or review</a>
-</axa-header-main-navigation>
+]'></axa-header-navigation>
