+++ src/components/m-header-navigation/index.scss
@@ -1,0 +1,62 @@
+@import "settings/header-navigation";
+@import "../a-link/mixins/link";
+@import "../../styles/mixins/respond";
+@import "../../styles/mixins/typo";
+@import "../../styles/mixins/unstyle-list";
+@import "styles/stroke";
+
+.m-header-navigation {
+  position: relative;
+  z-index: $header-navigation-z-index;
+
+  display: block;
+  height: $header-navigation-height;
+  border-bottom: $header-navigation-border-bottom-width solid $header-navigation-border-bottom-color;
+  background: $header-navigation-bg;
+}
+
+.m-header-navigation__nav {
+  display: block;
+
+  margin-left: auto;
+
+  @include respond-down(md) {
+    display: none;
+  }
+}
+
+.m-header-navigation__list {
+  @include unstyle-list();
+
+  display: flex;
+}
+
+.m-header-navigation__list-item {
+  display: block;
+
+  float: left;
+
+  margin: 0 $header-navigation__list-item-x-margin;
+}
+
+.m-header-navigation__list-link {
+  @include link-color();
+  @include typo-13--semibold();
+
+  position: relative;
+
+  display: flex;
+  height: $header-navigation-height;
+  padding: 0 $header-navigation__list-link-x-padding;
+
+  align-items: center;
+  text-align: center;
+  text-transform: uppercase;
+  vertical-align: middle;
+
+  &:hover,
+  &:active,
+  &:focus {
+    text-decoration: none;
+  }
+}
