+++ src/components/m-header-navigation/settings/_header-navigation.scss
@@ -1,0 +1,29 @@
+@import "../../../styles/settings/colors";
+@import "../../../styles/settings/zindex";
+
+$header-navigation-bg: $color-prim-white !default;
+$header-navigation-height: 75px !default;
+$header-navigation-z-index: $zindex-header-main-navigation !default;
+
+$header-navigation-border-bottom-color: $color-prim-gray-mercury !default;
+$header-navigation-border-bottom-width: 2px !default;
+
+$header-navigation__logo-size: 35px !default;
+$header-navigation__logo-size--md: 43px !default;
+$header-navigation__logo-size--xl: 53px !default;
+
+$header-navigation__list-item-x-margin: 8px !default;
+
+$header-navigation__list-link-x-padding: 10px !default;
+
+$header-navigation__form-y-padding: 6px !default;
+$header-navigation__form-padding-side: 20px !default;
+$header-navigation__form-border-side-width: 1px !default;
+$header-navigation__form-border-side-color: $color-prim-gray-silver !default;
+$header-navigation__form-margin-side: 20px !default;
+
+$header-navigation__icon-size: 25px !default;
+$header-navigation__search-icon-fill: $color-prim-gray !default;
+
+$header-navigation__burger-color: $color-axa-blue !default;
+$header-navigation__burger-margin-left: 15px !default;
