+++ src/components/m-header-navigation/styles/_stroke.scss
